Add umple models to the build chain, I'll use the same ones in the front end.
