+++ .gitignore
@@ -3,6 +3,7 @@ bin/
 lib/
 .project
 .classpath
+src/gen/
 
 *.class
