--- src/main/ump/DownloadSpec.ump
@@ -1,36 +1,0 @@
-namespace org.umple.sample.downloader;
-
-external FileType {}
-external URL {}
-
-/**
- * Specifies information about a download
- */
-interface DownloadSpec {
-  
-  /**
-   * Specify the base URL
-   */
-  URL getBaseName();
-  
-  FileType getFileType();
-  
-  List<URL> getResources();
-}
-
-class SpecUtil {
-
-  depend java.net.*;
-  
-  /**
-   * Convenience method to remove the {@code throws} concern from {@link URL#URL(String)} method. 
-   */
-  static URL createUrl(String path) {
-  	try {
-  	  return new URL(path);
-  	} catch (MalformedURLException mue) {
-  	  throw new IllegalArgumentException(mue);
-  	}
-  }
-
-}
